requirements.txt
asgiref==3.6.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
pytz==2023.3
requests==2.30.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
